Put things back
