--- com.unity.netcode.gameobjects/Runtime/SceneManagement/SceneMap.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 0df5a20fcb4a4cf0b9959485cee438c5
-timeCreated: 1761147915
